test(auth-commands): split login tests into logical modules
Split the ~970-line login/tests.rs into a shared test_support module (the
login_fake! fake, its scripted-prompt / read types, and the client / option /
prompt helpers) plus three scenario modules — test_non_interactive,
test_web_login, and test_classic_login. Pure code motion: no test body,
assertion, or fake logic changed.
